tss_api: impl reshare v2
