+++ backend/openapi/src/tss/request.ts
@@ -91,3 +91,42 @@ export const KeygenRequestV2Schema = registry.register(
     }),
   }),
 );
+
+export const ReshareRequestV2Schema = registry.register(
+  "TssUserReshareRequestV2",
+  z.object({
+    wallets: z
+      .object({
+        secp256k1: z
+          .string()
+          .openapi({
+            description: "secp256k1 public key in hex format (33 bytes)",
+          })
+          .optional(),
+        ed25519: z
+          .string()
+          .openapi({
+            description: "ed25519 public key in hex format (32 bytes)",
+          })
+          .optional(),
+      })
+      .refine((data) => data.secp256k1 || data.ed25519, {
+        message: "At least one of secp256k1 or ed25519 must be provided",
+      })
+      .openapi({
+        description: "Wallet public keys",
+      }),
+    reshared_key_shares: z
+      .array(
+        z.object({
+          name: z.string().openapi({
+            description: "Key share node name",
+          }),
+          endpoint: z.string().openapi({
+            description: "Key share node endpoint",
+          }),
+        }),
+      )
+      .openapi({ description: "Reshared key shares grouped by node" }),
+  }),
+);
