+++ backend/tss_api/src/api/v2/user/index.ts
@@ -25,10 +25,10 @@ import type {
 import { getKeyShareNodeMeta } from "@oko-wallet/oko-pg-interface/key_share_node_meta";
 import type { Wallet } from "@oko-wallet/oko-types/wallets";
 import type { NodeNameAndEndpoint } from "@oko-wallet/oko-types/user_key_share";
-import type { Bytes33 } from "@oko-wallet/bytes";
+import type { Bytes32, Bytes33 } from "@oko-wallet/bytes";
 
 import { generateUserTokenV2 } from "@oko-wallet-tss-api/api/keplr_auth";
-import { checkKeyShareFromKSNodes } from "@oko-wallet-tss-api/api/ks_node";
+import { checkKeyShareFromKSNodesV2 } from "@oko-wallet-tss-api/api/ks_node";
 
 export async function signInV2(
   db: Pool,
@@ -133,7 +133,7 @@ export async function signInV2(
     return {
       success: false,
       code: "UNKNOWN_ERROR",
-      msg: `signIn error: ${error instanceof Error ? error.message : String(error)}`,
+      msg: `signInV2 error: ${error instanceof Error ? error.message : String(error)}`,
     };
   }
 }
@@ -293,14 +293,25 @@ export async function checkEmail(
   }
 }
 
-export async function updateWalletKSNodesForReshare(
+export async function updateWalletKSNodesForReshareV2(
   db: Pool,
   email: string,
   auth_type: AuthType,
-  public_key: Bytes33,
+  wallets: {
+    secp256k1?: Bytes33;
+    ed25519?: Bytes32;
+  },
   reshared_key_shares: NodeNameAndEndpoint[],
 ): Promise<OkoApiResponse<void>> {
   try {
+    if (!wallets.secp256k1 && !wallets.ed25519) {
+      return {
+        success: false,
+        code: "INVALID_REQUEST",
+        msg: "At least one wallet (secp256k1 or ed25519) must be provided",
+      };
+    }
+
     const getUserRes = await getUserByEmailAndAuthType(db, email, auth_type);
     if (getUserRes.success === false) {
       return {
@@ -326,42 +337,6 @@ export async function updateWalletKSNodesForReshare(
       };
     }
 
-    const walletRes = await getWalletByPublicKey(
-      db,
-      Buffer.from(public_key.toUint8Array()),
-    );
-    if (walletRes.success === false) {
-      return {
-        success: false,
-        code: "UNKNOWN_ERROR",
-        msg: `getActiveWalletByUserIdAndCurveType error: ${walletRes.err}`,
-      };
-    }
-    if (walletRes.data === null) {
-      return {
-        success: false,
-        code: "WALLET_NOT_FOUND",
-        msg: `Wallet not found: ${email}`,
-      };
-    }
-
-    const wallet: Wallet = walletRes.data;
-    if (wallet.status !== "ACTIVE") {
-      return {
-        success: false,
-        code: "FORBIDDEN",
-        msg: `Wallet is not active: ${wallet.public_key.toString("hex")}`,
-      };
-    }
-
-    if (wallet.user_id !== user.user_id) {
-      return {
-        success: false,
-        code: "FORBIDDEN",
-        msg: `wallet user_id mismatch: ${email}`,
-      };
-    }
-
     const serverUrls = Array.from(
       new Set(reshared_key_shares.map((n) => n.endpoint)),
     );
@@ -381,29 +356,135 @@ export async function updateWalletKSNodesForReshare(
       };
     }
 
-    const checkKeyshareFromKSNodesRes = await checkKeyShareFromKSNodes(
+    const checkKeyshareFromKSNodesRes = await checkKeyShareFromKSNodesV2(
       email,
-      public_key,
+      wallets,
       getKSNodesRes.data,
       auth_type,
-      wallet.curve_type,
     );
     if (checkKeyshareFromKSNodesRes.success === false) {
       return checkKeyshareFromKSNodesRes;
     }
 
-    const nodeIds = checkKeyshareFromKSNodesRes.data.nodeIds;
-    const upsertWalletKSNodesRes = await upsertWalletKSNodes(
-      db,
-      wallet.wallet_id,
-      nodeIds,
-    );
-    if (upsertWalletKSNodesRes.success === false) {
-      return {
-        success: false,
-        code: "UNKNOWN_ERROR",
-        msg: `upsertWalletKSNodes error: ${upsertWalletKSNodesRes.err}`,
-      };
+    // Update wallet KS nodes for each wallet type
+    if (wallets.secp256k1 && checkKeyshareFromKSNodesRes.data.secp256k1) {
+      const walletRes = await getWalletByPublicKey(
+        db,
+        Buffer.from(wallets.secp256k1.toUint8Array()),
+      );
+      if (walletRes.success === false) {
+        return {
+          success: false,
+          code: "UNKNOWN_ERROR",
+          msg: `getWalletByPublicKey (secp256k1) error: ${walletRes.err}`,
+        };
+      }
+      if (walletRes.data === null) {
+        return {
+          success: false,
+          code: "WALLET_NOT_FOUND",
+          msg: `Secp256k1 wallet not found: ${email}`,
+        };
+      }
+
+      const wallet: Wallet = walletRes.data;
+      if (wallet.status !== "ACTIVE") {
+        return {
+          success: false,
+          code: "FORBIDDEN",
+          msg: `Secp256k1 wallet is not active: ${wallet.public_key.toString("hex")}`,
+        };
+      }
+
+      if (wallet.user_id !== user.user_id) {
+        return {
+          success: false,
+          code: "FORBIDDEN",
+          msg: `Secp256k1 wallet user_id mismatch: ${email}`,
+        };
+      }
+
+      if (wallet.curve_type !== "secp256k1") {
+        return {
+          success: false,
+          code: "FORBIDDEN",
+          msg: `Wallet curve type mismatch: expected secp256k1, got ${wallet.curve_type}`,
+        };
+      }
+
+      const nodeIds = checkKeyshareFromKSNodesRes.data.secp256k1.nodeIds;
+      const upsertWalletKSNodesRes = await upsertWalletKSNodes(
+        db,
+        wallet.wallet_id,
+        nodeIds,
+      );
+      if (upsertWalletKSNodesRes.success === false) {
+        return {
+          success: false,
+          code: "UNKNOWN_ERROR",
+          msg: `upsertWalletKSNodes (secp256k1) error: ${upsertWalletKSNodesRes.err}`,
+        };
+      }
+    }
+
+    if (wallets.ed25519 && checkKeyshareFromKSNodesRes.data.ed25519) {
+      const walletRes = await getWalletByPublicKey(
+        db,
+        Buffer.from(wallets.ed25519.toUint8Array()),
+      );
+      if (walletRes.success === false) {
+        return {
+          success: false,
+          code: "UNKNOWN_ERROR",
+          msg: `getWalletByPublicKey (ed25519) error: ${walletRes.err}`,
+        };
+      }
+      if (walletRes.data === null) {
+        return {
+          success: false,
+          code: "WALLET_NOT_FOUND",
+          msg: `Ed25519 wallet not found: ${email}`,
+        };
+      }
+
+      const wallet: Wallet = walletRes.data;
+      if (wallet.status !== "ACTIVE") {
+        return {
+          success: false,
+          code: "FORBIDDEN",
+          msg: `Ed25519 wallet is not active: ${wallet.public_key.toString("hex")}`,
+        };
+      }
+
+      if (wallet.user_id !== user.user_id) {
+        return {
+          success: false,
+          code: "FORBIDDEN",
+          msg: `Ed25519 wallet user_id mismatch: ${email}`,
+        };
+      }
+
+      if (wallet.curve_type !== "ed25519") {
+        return {
+          success: false,
+          code: "FORBIDDEN",
+          msg: `Wallet curve type mismatch: expected ed25519, got ${wallet.curve_type}`,
+        };
+      }
+
+      const nodeIds = checkKeyshareFromKSNodesRes.data.ed25519.nodeIds;
+      const upsertWalletKSNodesRes = await upsertWalletKSNodes(
+        db,
+        wallet.wallet_id,
+        nodeIds,
+      );
+      if (upsertWalletKSNodesRes.success === false) {
+        return {
+          success: false,
+          code: "UNKNOWN_ERROR",
+          msg: `upsertWalletKSNodes (ed25519) error: ${upsertWalletKSNodesRes.err}`,
+        };
+      }
     }
 
     return {
