+++ backend/tss_api/src/routes/v1/index.ts
@@ -5,7 +5,6 @@ import { setTriplesV1Routes } from "./triples";
 import { setPresignV1Routes } from "./presign";
 import { setSignV1Routes } from "./sign";
 import { setUserV1Routes } from "./user";
-import { setWalletEd25519Routes } from "./wallet_ed25519";
 import { setTssSessionRoutes } from "./tss_session";
 import { setKSNodeTelemetryRoutes } from "./ks_node_telemetry";
 
@@ -18,7 +17,6 @@ export function makeV1Router() {
   setSignV1Routes(router);
   setUserV1Routes(router);
 
-  setWalletEd25519Routes(router);
   setTssSessionRoutes(router);
   setKSNodeTelemetryRoutes(router);
