+++ backend/tss_api/src/routes/v2/index.ts
@@ -5,8 +5,13 @@ import { tssActivateMiddleware } from "@oko-wallet-tss-api/middleware/tss_activa
 import { keygenV2 } from "./keygen";
 import { userJwtMiddlewareV2 } from "@oko-wallet-tss-api/middleware/keplr_auth";
 import { presignStep1 } from "./presign_step_1";
+import { presignStep2 } from "./presign_step_2";
+import { presignStep3 } from "./presign_step_3";
 import { signStep1 } from "./sign_step_1";
 import { signStep2 } from "./sign_step_2";
+import { signEd25519Round1 } from "./sign_ed25519_round1";
+import { signEd25519Round2 } from "./sign_ed25519_round2";
+import { signEd25519Aggregate } from "./sign_ed25519_aggregate";
 import { apiKeyMiddleware } from "@oko-wallet-tss-api/middleware/api_key_auth";
 import { triplesStep1 } from "./triples_step_1";
 import { triplesStep2 } from "./triples_step_2";
@@ -20,12 +25,13 @@ import { triplesStep9 } from "./triples_step_9";
 import { triplesStep10 } from "./triples_step_10";
 import { triplesStep11 } from "./triples_step_11";
 import { keygenEd25519 } from "./keygen_ed25519";
-import { userSignin } from "./user_signin";
+import { userSignInV2 } from "./user_signin";
+import { userReshareV2 } from "./user_reshare";
 
 export function makeV2Router() {
   const router = Router();
 
-  router.post("/keygen/v2", oauthMiddleware, tssActivateMiddleware, keygenV2);
+  router.post("/keygen", oauthMiddleware, tssActivateMiddleware, keygenV2);
 
   router.post(
     "/keygen_ed25519",
@@ -41,6 +47,18 @@ export function makeV2Router() {
   );
 
   router.post(
+    "/presign/step2",
+    [userJwtMiddlewareV2, tssActivateMiddleware],
+    presignStep2,
+  );
+
+  router.post(
+    "/presign/step3",
+    [userJwtMiddlewareV2, tssActivateMiddleware],
+    presignStep3,
+  );
+
+  router.post(
     "/sign/step1",
     [userJwtMiddlewareV2, tssActivateMiddleware],
     signStep1,
@@ -50,6 +68,24 @@ export function makeV2Router() {
     "/sign/step2",
     [userJwtMiddlewareV2, tssActivateMiddleware],
     signStep2,
+  );
+
+  router.post(
+    "/sign_ed25519/round1",
+    [apiKeyMiddleware, userJwtMiddlewareV2, tssActivateMiddleware],
+    signEd25519Round1,
+  );
+
+  router.post(
+    "/sign_ed25519/round2",
+    [userJwtMiddlewareV2, tssActivateMiddleware],
+    signEd25519Round2,
+  );
+
+  router.post(
+    "/sign_ed25519/aggregate",
+    [userJwtMiddlewareV2, tssActivateMiddleware],
+    signEd25519Aggregate,
   );
 
   router.post(
@@ -122,7 +158,14 @@ export function makeV2Router() {
     "/user/signin",
     oauthMiddleware,
     tssActivateMiddleware,
-    userSignin,
+    userSignInV2,
+  );
+
+  router.post(
+    "/user/reshare",
+    oauthMiddleware,
+    tssActivateMiddleware,
+    userReshareV2,
   );
 
   return router;
