+++ backend/tss_api/src/routes/v2/sign_ed25519_round2.ts
@@ -1,0 +1,119 @@
+import type { Response } from "express";
+import type {
+  SignEd25519Round2Body,
+  SignEd25519Round2Response,
+} from "@oko-wallet/oko-types/tss";
+import type { OkoApiResponse } from "@oko-wallet/oko-types/api_response";
+import { ErrorCodeMap } from "@oko-wallet/oko-api-error-codes";
+import {
+  ErrorResponseSchema,
+  UserAuthHeaderSchema,
+} from "@oko-wallet/oko-api-openapi/common";
+import { registry } from "@oko-wallet/oko-api-openapi";
+
+import { runSignEd25519Round2 } from "@oko-wallet-tss-api/api/sign_ed25519";
+import {
+  type UserAuthenticatedRequest,
+  sendResponseWithNewToken,
+} from "@oko-wallet-tss-api/middleware/keplr_auth";
+
+registry.registerPath({
+  method: "post",
+  path: "/tss/v2/sign_ed25519/round2",
+  tags: ["TSS"],
+  summary: "Generate Ed25519 signature share (Round 2)",
+  description: "Server generates signature share using collected commitments",
+  security: [{ userAuth: [] }],
+  request: {
+    headers: UserAuthHeaderSchema,
+    body: {
+      required: true,
+      content: {
+        "application/json": {
+          schema: {
+            type: "object",
+            properties: {
+              session_id: { type: "string" },
+              commitments_1: {
+                type: "object",
+                properties: {
+                  identifier: { type: "array", items: { type: "number" } },
+                  commitments: { type: "array", items: { type: "number" } },
+                },
+              },
+            },
+            required: ["session_id", "commitments_1"],
+          },
+        },
+      },
+    },
+  },
+  responses: {
+    200: {
+      description: "Successfully generated signature share",
+      content: {
+        "application/json": {
+          schema: {
+            type: "object",
+            properties: {
+              success: { type: "boolean" },
+              data: {
+                type: "object",
+                properties: {
+                  signature_share_0: {
+                    type: "object",
+                    properties: {
+                      identifier: {
+                        type: "array",
+                        items: { type: "number" },
+                      },
+                      signature_share: {
+                        type: "array",
+                        items: { type: "number" },
+                      },
+                    },
+                  },
+                },
+              },
+            },
+          },
+        },
+      },
+    },
+    401: {
+      description: "Unauthorized",
+      content: { "application/json": { schema: ErrorResponseSchema } },
+    },
+    500: {
+      description: "Internal server error",
+      content: { "application/json": { schema: ErrorResponseSchema } },
+    },
+  },
+});
+
+export async function signEd25519Round2(
+  req: UserAuthenticatedRequest<SignEd25519Round2Body>,
+  res: Response<OkoApiResponse<SignEd25519Round2Response>>,
+) {
+  const state = req.app.locals as any;
+  const user = res.locals.user;
+  const body = req.body;
+
+  const result = await runSignEd25519Round2(
+    state.db,
+    state.encryption_secret,
+    {
+      email: user.email.toLowerCase(),
+      wallet_id: user.wallet_id_ed25519,
+      session_id: body.session_id,
+      commitments_1: body.commitments_1,
+    },
+  );
+
+  if (result.success === false) {
+    res.status(ErrorCodeMap[result.code] ?? 500).json(result);
+    return;
+  }
+
+  sendResponseWithNewToken(res, result.data);
+}
